requirements.txt
# Core AI & LLM Dependencies
langchain-groq
langchain-ollama
langchain-core

# Audio Processing
edge-tts
pygame
faster-whisper
pyaudio
openwakeword  # Efficient wake word detection

# System Control & Automation
psutil
pyautogui
keyboard
requests
pywin32

# Image Processing
pillow
opencv-python
mss

# GUI Framework
PyQt6

# Environment & Utilities
python-dotenv
colorama
pyperclip

# Security - Encryption
cryptography

# Vector Memory & Semantic Search (Optional - for advanced features)
chromadb
sentence-transformers
sounddevice  # For wake word training
